refactor(scripts): move ingestion/platform to ingestion/alert_api (#175)
