+++ annotation_api/scripts/data_transfer/ingestion/alert_api/progress_management.py
@@ -286,10 +286,10 @@ class LogSuppressor:
                 "",  # root logger - most important
                 "__main__",
                 "root",
-                "scripts.data_transfer.ingestion.platform.import",
-                "scripts.data_transfer.ingestion.platform.shared",
-                "scripts.data_transfer.ingestion.platform.client",
-                "scripts.data_transfer.ingestion.platform.utils",
+                "scripts.data_transfer.ingestion.alert_api.import",
+                "scripts.data_transfer.ingestion.alert_api.shared",
+                "scripts.data_transfer.ingestion.alert_api.client",
+                "scripts.data_transfer.ingestion.alert_api.utils",
                 "app.clients.annotation_api",
                 "requests",
                 "urllib3",
